requirements.txt
filelock==3.14.0
fonttools==4.51.0
idna==2.6
image==1.5.33
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
soundfile==0.12.1
tkinter==8.6.13
urllib3==2.2.1
virtualenv==20.26.2
